refactoring: little changes in folders structure
