+++ src/pages/ColorDetailsPage/ColorDetailsPage.jsx
