+++ src/pages/ColorDetailsPage/ColorDetailsPage.module.scss
