+++ src/pages/ColorsPage/ColorsPage.jsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from 'react';
 import styles from './ColorsPage.module.scss';
 import axios from 'axios';
 import { useParams, useNavigate } from 'react-router-dom';
-import ColorCard from '../ColorCard/ColorCard';
-import Pagination from '../Pagination/Pagination';
+import ColorCard from '../../components/ColorCard/ColorCard';
+import Pagination from '../../components/Pagination/Pagination';
 import { useSelector } from 'react-redux';
 
 const ColorsPage = () => {
@@ -65,13 +65,15 @@ const ColorsPage = () => {
         {renderColors}
       </div>
 
-      <Pagination
-        activePage={routeParams.page}
-        totalPages={totalPages}
-        onBtnNumber={handleChosenPage}
-        onIncreasePage={increasePageNumber}
-        onDecreasePageNumber={decreasePageNumber}
-      />
+      <div className={styles['colors-page__pagination']}>
+        <Pagination
+          activePage={routeParams.page}
+          totalPages={totalPages}
+          onBtnNumber={handleChosenPage}
+          onIncreasePage={increasePageNumber}
+          onDecreasePageNumber={decreasePageNumber}
+        />
+      </div>
     </div>
   )
 }
