+++ src/pages/People-page/People-page.module.scss
