+++ src/redux/peopleReducer.js
@@ -1,4 +1,4 @@
-import { INIT_LIST_OF_PEOPLE, ADD_NEW_PERSON, DELETE_PERSON } from './types';
+import { INIT_LIST_OF_PEOPLE, ADD_NEW_PERSON, DELETE_PERSON, CLEAR_LIST_OF_PEOPLE } from './types';
 
 const peopleCollection = {
   people: [],
@@ -8,6 +8,8 @@ export const peopleReducer = (state = peopleCollection, action) => {
   switch (action.type) {
     case INIT_LIST_OF_PEOPLE:
       return {...state, people: state.people.concat(action.payload)};
+    case CLEAR_LIST_OF_PEOPLE:
+      return {...state, people: []};
     case ADD_NEW_PERSON:
       const isAlreadyAdded = state.people.find(person => person.id === action.payload.id);
       if (!isAlreadyAdded) {
